!68 resolve quick sort array out of bounds problem * use signed int in quicksort to resolve array out of bounds * code refactor for order by
Former-commit-id: 452e713be959bc5142a2beeb5b54bafebd26fd25

+++ joy/cpp/src/jni/sort_api.h
@@ -4,32 +4,32 @@
 #include "../operator/sort.h"
 
 // return the jit context address
-long sortPrepare(
-    int *sourceTypes,
-    int typeCount,
-    int *outputCols,
-    int outputColCount,
-    int *sortCols,
-    int *sortAscendings,
-    int *sortNullFirsts,
-    int sortColCount);
+int64_t sortPrepare(
+    int32_t *sourceTypes,
+    int32_t typeCount,
+    int32_t *outputCols,
+    int32_t outputColCount,
+    int32_t *sortCols,
+    int32_t *sortAscendings,
+    int32_t *sortNullFirsts,
+    int32_t sortColCount);
 
 // return the sort operator address
-long sortCreateOperator(
-    long contextAddress,
-    int *sourceTypes,
-    int typeCount,
-    int *outputCols,
-    int outputColCount,
-    int *sortCols,
-    int *sortAscendings,
-    int *sortNullFirsts,
-    int sortColCount);
+int64_t sortCreateOperator(
+    int64_t contextAddress,
+    int32_t *sourceTypes,
+    int32_t typeCount,
+    int32_t *outputCols,
+    int32_t outputColCount,
+    int32_t *sortCols,
+    int32_t *sortAscendings,
+    int32_t *sortNullFirsts,
+    int32_t sortColCount);
 
-void sortAddInput(long contextAddress, long sortAddress, long *datas, long *nulls, int pageCount, long *rowCounts, long totalRowCount);
+void sortAddInput(int64_t contextAddress, int64_t sortAddress, int64_t *datas, int64_t *nulls, int32_t pageCount, int32_t *rowCounts, int32_t totalRowCount);
 
-void sortExecute(long contextAddress, long sortAddress);
+void sortExecute(int64_t contextAddress, int64_t sortAddress);
 
-Table *sortGetOutput(long contextAddress, long sortAddress);
+Table *sortGetOutput(int64_t contextAddress, int64_t sortAddress);
 
 #endif
